Open-LLM-VTuber/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.16
    # via edge-tts
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anthropic==0.49.0
    # via open-llm-vtuber (pyproject.toml)
anyio==4.9.0
    # via
    #   anthropic
    #   groq
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
async-timeout==5.0.1
    # via aiohttp
attrs==25.3.0
    # via aiohttp
azure-cognitiveservices-speech==1.43.0
    # via open-llm-vtuber (pyproject.toml)
certifi==2025.1.31
    # via
    #   edge-tts
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via soundfile
cfgv==3.4.0
    # via pre-commit
chardet==5.2.0
    # via open-llm-vtuber (pyproject.toml)
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
coloredlogs==15.0.1
    # via onnxruntime
distlib==0.3.9
    # via virtualenv
distro==1.9.0
    # via
    #   anthropic
    #   groq
    #   openai
dnspython==2.7.0
    # via email-validator
edge-tts==7.0.0
    # via open-llm-vtuber (pyproject.toml)
email-validator==2.2.0
    # via fastapi
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.12
    # via open-llm-vtuber (pyproject.toml)
fastapi-cli==0.0.7
    # via fastapi
filelock==3.18.0
    # via
    #   torch
    #   virtualenv
flatbuffers==25.2.10
    # via onnxruntime
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.3.2
    # via torch
groq==0.22.0
    # via open-llm-vtuber (pyproject.toml)
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   open-llm-vtuber (pyproject.toml)
    #   anthropic
    #   fastapi
    #   groq
    #   letta-client
    #   openai
httpx-sse==0.4.0
    # via letta-client
humanfriendly==10.0
    # via coloredlogs
identify==2.6.9
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.6
    # via
    #   fastapi
    #   torch
jiter==0.9.0
    # via
    #   anthropic
    #   openai
langdetect==1.0.9
    # via open-llm-vtuber (pyproject.toml)
letta-client==0.1.101
    # via open-llm-vtuber (pyproject.toml)
loguru==0.7.3
    # via open-llm-vtuber (pyproject.toml)
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.2.0
    # via
    #   aiohttp
    #   yarl
networkx==3.4.2
    # via torch
nodeenv==1.9.1
    # via pre-commit
numpy==1.26.4
    # via
    #   open-llm-vtuber (pyproject.toml)
    #   onnxruntime
    #   scipy
    #   soundfile
onnxruntime==1.21.0
    # via
    #   open-llm-vtuber (pyproject.toml)
    #   silero-vad
openai==1.72.0
    # via open-llm-vtuber (pyproject.toml)
packaging==24.2
    # via onnxruntime
platformdirs==4.3.7
    # via virtualenv
pre-commit==4.2.0
    # via open-llm-vtuber (pyproject.toml)
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
protobuf==6.30.2
    # via onnxruntime
pycparser==2.22
    # via cffi
pydantic==2.11.3
    # via
    #   anthropic
    #   fastapi
    #   groq
    #   letta-client
    #   openai
pydantic-core==2.33.1
    # via
    #   letta-client
    #   pydantic
pydub==0.25.1
    # via open-llm-vtuber (pyproject.toml)
pygments==2.19.1
    # via rich
    # via pyttsx3
    # via
     pysbd==0.3.4
    # via open-llm-vtuber (pyproject.toml)
python-dotenv==1.1.0
    # via uvicorn
python-multipart==0.0.20
    # via fastapi
pyttsx3==2.98
    # via open-llm-vtuber (pyproject.toml)
pyyaml==6.0.2
    # via
    #   open-llm-vtuber (pyproject.toml)
    #   pre-commit
    #   uvicorn
requests==2.32.3
    # via open-llm-vtuber (pyproject.toml)
rich==14.0.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.14.1
    # via fastapi-cli
ruamel-yaml==0.18.10
    # via open-llm-vtuber (pyproject.toml)
ruamel-yaml-clib==0.2.12
    # via ruamel-yaml
ruff==0.11.4
    # via open-llm-vtuber (pyproject.toml)
scipy==1.15.2
    # via open-llm-vtuber (pyproject.toml)
shellingham==1.5.4
    # via typer
sherpa-onnx==1.12.25
    # via open-llm-vtuber (pyproject.toml)
silero-vad==5.1.2
    # via open-llm-vtuber (pyproject.toml)
six==1.17.0
    # via langdetect
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   groq
    #   openai
soundfile==0.13.1
    # via open-llm-vtuber (pyproject.toml)
srt==3.5.3
    # via edge-tts
starlette==0.46.1
    # via fastapi
sympy==1.13.1
    # via
    #   onnxruntime
    #   torch
tabulate==0.9.0
    # via edge-tts
tomli==2.2.1
    # via open-llm-vtuber (pyproject.toml)
torch==2.6.0
    # via
    #   open-llm-vtuber (pyproject.toml)
    #   silero-vad
    #   torchaudio
torchaudio==2.6.0
    # via silero-vad
tqdm==4.67.1
    # via
    #   open-llm-vtuber (pyproject.toml)
    #   openai
typer==0.15.2
    # via fastapi-cli
typing-extensions==4.13.1
    # via
    #   anthropic
    #   anyio
    #   edge-tts
    #   fastapi
    #   groq
    #   letta-client
    #   multidict
    #   openai
    #   pydantic
    #   pydantic-core
    #   rich
    #   rich-toolkit
    #   torch
    #   typer
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.0
    # via pydantic
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via
    #   open-llm-vtuber (pyproject.toml)
    #   fastapi
    #   fastapi-cli
    # via uvicorn
virtualenv==20.30.0
    # via pre-commit
watchfiles==1.0.5
    # via uvicorn
websocket-client==1.8.0
    # via open-llm-vtuber (pyproject.toml)
websockets==15.0.1
    # via uvicorn
yarl==1.19.0
    # via aiohttp
